Adjusting the way AsyncIO is calling out
