+++ requirements.txt
@@ -41,3 +41,4 @@ urllib3==1.21.1
 Werkzeug==0.12.2
 zope.interface==4.4.1
 python-jose==1.3.2
+treq==17.3.1
