+++ swagger_server/controllers_local/proxy_controller.py
@@ -7,32 +7,20 @@
 ##############################################################################
 from flask import jsonify, make_response
 from ras_api_gateway.proxy_router import router
-from twisted.internet import reactor
-from twisted.internet.error import ConnectionRefusedError, NoRouteError, UserError
-from twisted.web.client import Agent, readBody
-from twisted.internet.defer import DeferredList
-from crochet import wait_for
 from json import loads
 from jinja2 import Environment, FileSystemLoader
-from datetime import datetime, timedelta
 from ..configuration import ons_env
-import arrow
 import twisted.internet._sslverify as v
+from .aggregation import ONSAggregation
+from traceback import print_exc
+from sys import stdout
 #
 #   Disable SSL tail certificate verification
 #
 v.platformTrust = lambda : None
 
 env = Environment(loader=FileSystemLoader('templates'))
-
-inputDateFormat = 'YYYY-MM-DDThh:mm:ss'
-outputDateFormat = 'D MMM YYYY'
-
-CASES_GET = '/collection-exercise-api/1.0.0/cases/partyid'
-RESPONDENTS_GET = '/collection-exercise-api/1.0.0/respondents'
-SURVEY_GET = '/collection-exercise-api/1.0.0/surveys'
-BUSINESS_GET = '/collection-exercise-api/1.0.0/businesses/id'
-EXERCISE_GET = '/collection-exercise-api/1.0.0/collection-exercise'
+aggregator = ONSAggregation()
 
 
 def get_secret(*args, **kwargs):
@@ -91,42 +79,6 @@ def ping(host, port):
     return make_response(msg, code)
 
 
-def calculate_case_status(caseEvents):
-    status = ''
-    for event in caseEvents:
-        if event['category'] == 'CASE_UPLOADED':
-            status = 'Complete'
-            break
-    if status == '':
-        for event in caseEvents:
-            if event['category'] == 'CASE_DOWNLOADED':
-                status = 'In progress'
-                break
-    if status == '':
-        status = 'Not started'
-    return status
-
-
-def hit_read(response):
-    if response.code != 200:
-        raise UserError
-    return response
-
-
-def hit_route(uri, params):
-    route = router.route(uri)
-    if not route: raise NoRouteError
-    url = '{}://{}:{}{}/{}'.format(route.proto, route.host, route.port, uri, params).encode()
-    deferred = Agent(reactor).request(b'GET', url, None)
-    deferred.addCallback(hit_read).addCallback(readBody)
-    return deferred
-
-
-@wait_for(timeout=10)
-def lookup_cases(party_id):
-    return hit_route(CASES_GET, party_id)
-
-
 def survey_todo(id=None, status_filter=None):
     """
     Generate the TODO data for the mySurveys page. This involves making a number of cross-MS
@@ -148,66 +100,9 @@ def survey_todo(id=None, status_filter=None):
     :return: An aggregated record to be consumed by the mySurveys page
     """
     try:
-        return survey_todo_process(id, loads(status_filter))
+        return aggregator.survey_todo(id, loads(status_filter))
     except Exception as e:
-        print('Unexpected exception: ',e)
+        print('Unexpected exception: ', e)
+        print_exc(limit=20, file=stdout)
         return make_response('unexpected error', 500)
 
-
-def survey_todo_process(party_id, status_filter):
-    """Query various endpoints and aggregate result"""
-    if type(status_filter) != list:
-        return make_response('"status filter" needs to be a JSON format list of statuses', 500)
-
-    results = {}
-    rows = []
-    try:
-        cases = loads(lookup_cases(party_id).decode())
-    except ConnectionRefusedError:
-        return make_response('case service is not responding', 500)
-    except NoRouteError:
-        return make_response('no route to case service', 500)
-    except UserError:
-        return make_response('no such party id "{}"'.format(party_id), 404)
-    except Exception as e:
-        print("Error>", e)
-        return make_response('unknown error', 500)
-
-    def attach(blob, case_id, key):
-        results[case_id][key] = loads(blob.decode())
-        return True, None
-
-    @wait_for(timeout=10)
-    def fetch_rows():
-        dlist = [hit_route(RESPONDENTS_GET, party_id)]
-        for case in cases:
-            def attach_exercise(ex, case_id):
-                ex = results[case_id]['exercise'] = loads(ex.decode())
-                for key in ['periodStart', 'periodEnd', 'scheduledReturn']:
-                    ex[key] = ex[key].replace('Z', '')
-                    ex[key + 'Formatted'] = arrow.get(ex[key], inputDateFormat).format(outputDateFormat)
-                return DeferredList([hit_route(SURVEY_GET, ex['surveyId']).addCallback(attach, case_id, 'survey')])
-            case_id = case['id']
-            business_id = case['caseGroup']['partyId']
-            exercise_id = case['caseGroup']['collectionExerciseId']
-            results[case_id] = {'case': case}
-            business = hit_route(BUSINESS_GET, business_id).addCallback(attach, case_id, 'business')
-            exercise = hit_route(EXERCISE_GET, exercise_id).addCallback(attach_exercise, case_id)
-            dlist += [business, exercise]
-        return DeferredList(dlist)
-
-    deferreds = fetch_rows()
-    for deferred in deferreds:
-        if not deferred[0]:
-            return deferred[1] if deferred[1] == str else deferred[1].getErrorMessage(), 500
-    for item in results.values():
-        item_status = calculate_case_status(item['case']['caseEvents']).lower()
-        if item_status in status_filter:
-            rows.append({
-                'businessData': item['business'],
-                'case': item['case'],
-                'collectionExerciseData': item['exercise'],
-                'surveyData': item['survey'],
-                'status': item_status
-            })
-    return make_response(jsonify({'userData': loads(deferreds[0][1].decode()), 'rows': rows}), 200)
